Updates to GUI functionality including transaction insertion and editing.
Created sql file to reset transactions to 10 items.
Added lots of code to pbsql.LedgerView that handles the table updates
Added hooks in gui.py to handle updates.
Played with using TreeCtrl for AccountList in gui.py, but decided to instead add my own 'selection highlighting' to the FoldPanelBar of AccountList.
Ledger tab instead of Summary now shows at program startup.
Added _data_update() to LedgerGridBaseTable - it's called a lot, but is working for now. Likely a memory/speed issue in the future.
Added a whole bunch of debug code from douglib so that I can time function calls.