More plotting work.
Added Pareto plot to plots.py along with demo. Majority of the code was taken from Abraham Lee's paretochart project (https://github.com/tisimst/paretochart), but I modified it so that it works with Python 3.4 (without needed to run 2to3) and so that it works with the wxPython matplotlib backend.
Switched .py file to build into exe back to gui.py.